pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fenn"
version = "0.0.2"
description = "Friendly Environment for Neural Networks – a simple framework that automates the boring side of ML."

authors = [
  { name="Alessio Russo", email="alessio.russo.ar98@gmail.com" },
]

readme = "README.md"

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]

dependencies = [
  "PyYAML>=6.0.3",
  "wandb>=0.23.0",
  "numpy>=2,<2.3.0",
  "pandas>=2.3.3",
  "matplotlib>=3.10.7",
  "scikit-learn>=1.7.2",
  "colorama>=0.4.6",
  "requests>=2.32.5",
  "python-dotenv>=1.2.1"
]

requires-python = ">=3.11"

[project.optional-dependencies]
torch = [
  "torch>=2.9.1",
  "torchvision>=0.24.1"
]
test = [
  "pytest>=8.0.0",
  "requests-mock>=1.11.0",
  "Faker>=25.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
fenn = "fenn.cli:main"

[project.urls]
"Homepage" = "https://fennpy.org"
"Documentation" = "https://fennpy.org"
"Source Code" = "https://github.com/blkdmr/fennpy"
"Bug Tracker" = "https://github.com/blkdmr/fennpy/issues"